Fix db
