+++ server.log
@@ -116,3 +116,100 @@ GET /api/contacts 15-01-2022_08:23:37
 GET /api/contacts 15-01-2022_08:24:48
 GET /api/contacts 15-01-2022_08:24:49
 GET /api/contacts 15-01-2022_08:24:50
+GET /api/contacts 16-01-2022_06:39:08
+GET /api/contacts/1 16-01-2022_06:39:48
+POST /api/contacts 16-01-2022_06:41:55
+POST /api/contacts 16-01-2022_06:43:25
+POST /api/contacts 16-01-2022_07:05:39
+POST /api/contacts 16-01-2022_07:06:09
+GET /api/contacts 16-01-2022_07:06:20
+PUT /api/contacts/1 16-01-2022_07:06:43
+GET /api/contacts/1 16-01-2022_07:06:54
+GET /api/contacts 16-01-2022_07:06:58
+DELETE /api/contacts/4 16-01-2022_07:07:19
+GET /api/contacts/4 16-01-2022_07:07:55
+GET /api/contacts 16-01-2022_07:08:03
+DELETE /api/contacts/4 16-01-2022_07:08:19
+DELETE /api/contacts/4 16-01-2022_07:08:20
+DELETE /api/contacts/4 16-01-2022_07:09:39
+GET /api/contacts 16-01-2022_07:12:13
+DELETE /api/contacts/4 16-01-2022_07:12:40
+DELETE /api/contacts/4 16-01-2022_07:14:09
+GET /api/contacts 16-01-2022_07:14:50
+GET /api/contacts 16-01-2022_07:14:58
+GET /api/contacts 16-01-2022_07:15:17
+GET /api/contacts/2 16-01-2022_07:15:21
+GET /api/contacts 16-01-2022_07:16:53
+GET /api/contacts/1 16-01-2022_07:17:03
+GET /api/contacts/2 16-01-2022_07:17:07
+GET /api/contacts/4 16-01-2022_07:17:10
+POST /api/contacts 16-01-2022_07:17:20
+GET /api/contacts 16-01-2022_07:17:26
+DELETE /api/contacts/1 16-01-2022_07:17:50
+DELETE /api/contacts 16-01-2022_07:18:02
+DELETE /api/contacts/1 16-01-2022_07:19:19
+GET /api/contacts 16-01-2022_07:19:31
+POST /api/contacts 16-01-2022_07:19:42
+GET /api/contacts 16-01-2022_07:19:47
+DELETE /api/contacts/4 16-01-2022_07:19:59
+GET /api/contacts 16-01-2022_07:21:35
+GET /api/contacts 16-01-2022_07:21:45
+GET /api/contacts 16-01-2022_07:21:46
+GET /api/contacts 16-01-2022_07:21:46
+GET /api/contacts 16-01-2022_07:23:48
+GET /api/contacts 16-01-2022_07:23:49
+GET /api/contacts 16-01-2022_07:24:10
+GET /api/contacts 16-01-2022_07:24:22
+GET /api/contacts 16-01-2022_07:24:52
+GET /api/contacts/1 16-01-2022_07:28:12
+GET /api/contacts/1 16-01-2022_07:28:40
+GET /api/contacts/6 16-01-2022_07:28:48
+GET /api/contacts 16-01-2022_07:28:51
+GET /api/contacts/2 16-01-2022_07:28:57
+GET /api/contacts 16-01-2022_07:29:00
+POST /api/contacts 16-01-2022_07:35:19
+GET /api/contacts 16-01-2022_07:35:39
+POST /api/contacts 16-01-2022_07:36:03
+POST /api/contacts 16-01-2022_07:36:12
+GET /api/contacts 16-01-2022_07:36:19
+GET /api/contacts 16-01-2022_07:42:28
+DELETE /api/contacts/748d894f-1245-498e-8c0e-73634d2628ad 16-01-2022_07:42:47
+GET /api/contacts 16-01-2022_07:43:05
+GET /api/contacts 16-01-2022_07:47:52
+PUT /api/contacts/499d108b-5385-4cf0-91d3-a481d8665b0b 16-01-2022_07:48:53
+GET /api/contacts 16-01-2022_07:49:13
+GET /api/contacts 16-01-2022_08:04:54
+GET /api/contacts 16-01-2022_08:08:02
+GET /api/contacts 16-01-2022_08:10:23
+GET /api/contacts 16-01-2022_08:15:15
+GET /api/contacts 16-01-2022_08:34:41
+GET /api/contacts/1 16-01-2022_08:34:48
+POST /api/contacts 16-01-2022_08:35:00
+GET /api/contacts 16-01-2022_08:35:06
+DELETE /api/contacts/741acafc-07ea-4fe6-a56f-9ec0ae254f60 16-01-2022_08:35:28
+GET /api/contacts 16-01-2022_08:35:39
+GET /api/contacts 16-01-2022_08:36:46
+GET /api/contacts/1 16-01-2022_08:36:54
+GET /api/contacts/2 16-01-2022_08:36:57
+GET /api/contacts 16-01-2022_08:36:59
+POST /api/contacts 16-01-2022_08:37:14
+POST /api/contacts 16-01-2022_08:38:41
+POST /api/contacts 16-01-2022_08:40:31
+POST /api/contacts 16-01-2022_08:41:43
+POST /api/contacts 16-01-2022_08:42:25
+POST /api/contacts 16-01-2022_08:42:50
+POST /api/contacts 16-01-2022_08:42:53
+POST /api/contacts 16-01-2022_08:44:08
+GET /api/contacts 16-01-2022_08:44:18
+GET /api/contacts 16-01-2022_08:45:43
+POST /api/contacts 16-01-2022_08:45:52
+POST /api/contacts 16-01-2022_08:46:18
+POST /api/contacts 16-01-2022_08:46:39
+POST /api/contacts 16-01-2022_08:47:02
+POST /api/contacts 16-01-2022_08:48:10
+POST /api/contacts 16-01-2022_08:48:23
+POST /api/contacts 16-01-2022_08:48:56
+POST /api/contacts 16-01-2022_08:49:41
+POST /api/contacts 16-01-2022_08:50:07
+POST /api/contacts 16-01-2022_08:51:56
+POST /api/contacts 16-01-2022_08:52:37
